backend/my_apps/users/pyproject.toml
[tool.poetry]
name = "mydrffinalprj"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
g4f = "^0.1.8.6"
djangorestframework = "^3.14.0"
isort = "^5.12.0"
django-filter = "^23.3"
djangorestframework-simplejwt = "^5.3.0"
werkzeug = "^3.0.1"
scrapy = "^2.11.0"
scrapy-djangoitem = "^1.1.1"
ipython = "^8.17.2"
scrapy-selenium = "^0.0.7"
free-proxy = "^1.1.1"
django-extensions = "^3.2.3"
drf-api-logger = "^1.1.15"
django-silk = "^5.0.4"
psycopg2 = "^2.9.9"
pandas = "^2.1.3"
setuptools = "^69.0.2"
wheel = "^0.41.3"
playwright = "^1.40.0"
django = "^4.2.7"
python-dotenv = "^1.0.0"
fake-useragent = "^1.3.0"
selenium-wire = "^5.1.0"
scrapy-playwright = "^0.0.33"
grequests = "^0.7.0"
update = "^0.0.1"
dataclasses = "0.6"
whitenoise = "^6.6.0"
drf-yasg = "1.21.7"
celery = "^5.3.6"
redis = "^5.0.1"
django-celery-results = "^2.5.1"
django-celery-beat = "^2.5.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
